\name{A_example}
\alias{A_example}
\docType{data}
\title{A_example data}
\description{
A binary matrix of class \code{matrix} used as toy example in \code{getPM} and \code{getMutex} and \code{getMutexAB} and \code{getMutexGroup}.
}
\usage{data("A_example")}
\format{
  The format is:
 num [1:1000, 1:500] 0 0 0 0 0 1 0 0 0 1 ...
}
\examples{
data(A_example)
}
\keyword{datasets}
