\name{B_example}
\alias{B_example}
\docType{data}
\title{B_example data}
\description{
A binary matrix of class \code{matrix} used as toy example in \code{getPM} and \code{getMutex} and \code{getMutexAB} and \code{getMutexGroup}.
}
\usage{data("B_example")}
\format{
  The format is:
  
 int [1:1000, 1:500] 0 1 1 0 1 0 0 0 1 1 ...
}
\examples{
data(B_example)
}
\keyword{datasets}
