\name{PM_AMP_COAD}
\alias{PM_AMP_COAD}
\docType{data}
\title{PM_AMP_COAD data}
\description{
Probability matrix, with information of genes being amplified in samples in Colon Adenocarcinoma, created by AMP_COAD.rda applying \code{getPM} and used as real example and \code{getMutexAB}.
}
\usage{data("PM_AMP_COAD")}
\format{
  The format is:
  
 num [1:1000, 1:391] 0.118 0.118 0.118 0.118 0.114 ...
}
\examples{
data(PM_AMP_COAD)
}
\keyword{datasets}
