\name{PM_TCGA_COAD_AMP}
\alias{PM_TCGA_COAD_AMP}
\docType{data}
\title{PM_TCGA_COAD_AMP data}
\description{
Probability matrix, with information of genes being mutated in samples in Colon Adenocarcinoma, created from TCGA_COAD_AMP.rda by applying \code{getPM} and used as real example in \code{getMutexAB}.
}
\usage{data("PM_TCGA_COAD_AMP")}
\format{
  The format is:
  
 num [1:1000, 1:391] 0.97 0.958 0.958 0.945 0.889 ...
}
\examples{
data(PM_TCGA_COAD_AMP)
}
\keyword{datasets}
