% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{InternalFunctions}
\alias{InternalFunctions}
\alias{speedglm.wfit2}
\alias{expand.grid_fast}
\alias{get_os}
\alias{control_2}
\alias{is.sparse_2}
\title{Rediscover Internal Functions}
\usage{
speedglm.wfit2(
  y,
  X,
  intercept = TRUE,
  weights = NULL,
  row.chunk = NULL,
  family = gaussian(),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = NULL,
  acc = 1e-08,
  maxit = 25,
  k = 2,
  sparselim = 0.9,
  camp = 0.01,
  eigendec = TRUE,
  tol.values = 1e-07,
  tol.vectors = 1e-07,
  tol.solve = .Machine$double.eps,
  sparse = NULL,
  method = c("eigen", "Cholesky", "qr"),
  trace = FALSE,
  ...
)

expand.grid_fast(a, b)

get_os()

control_2(
  B,
  symmetric = TRUE,
  tol.values = 1e-07,
  tol.vectors = 1e-07,
  out.B = TRUE,
  method = c("eigen", "Cholesky")
)

is.sparse_2(X, sparselim = 0.9, camp = 0.05)
}
\value{
Internal outputs
}
\description{
Internal functions used by EventPointer in the different
steps of the algorithm
}
\keyword{internal}
